// frontend/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        MemberStatisticsResponse: {
            /** Draws */
            draws: number;
            /** Games Played */
            games_played: number;
            /** Losses */
            losses: number;
            /** Member Id */
            member_id: number;
            /** Nickname */
            nickname: string;
            /** Rank */
            rank: number | null;
            /** Rating */
            rating: number;
            /** Wins */
            wins: number;
        };
// ...
        RankingsResponse: {
            /** Has More */
            has_more: boolean;
            /** Items */
            items: components["schemas"]["MemberStatisticsResponse"][];
            /** Limit */
            limit: number;
            me: components["schemas"]["MemberStatisticsResponse"] | null;
            /** Offset */
            offset: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}